+++ CHANGELOG.md
@@ -75,6 +75,7 @@
 
 #### New Features
 - Added support for the `dtypes` parameter of `pd.get_dummies`
+- Added support for the `random_state` parameter of `DataFrame.sample` and `Series.sample`
 
 #### Improvements
 
@@ -148,6 +149,7 @@
 #### Bug Fixes
 
 - Fixed a bug where the row count was not getting cached in the ordered dataframe each time count_rows() is called.
+- Fixed bug where `Series.sample` and `DataFrame.sample` would allow setting `n` larger than the number of rows while `replace=False`.
 
 ## 1.40.0 (2025-10-02)